add hook useLogin and useContext
