+++ src/components/Auth/FormComponents/FormInput.jsx
@@ -1,8 +1,11 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { StylesContext } from '../Login/StylesContext';
 
 export const FormInput = ({ label, type, name, value, onChange }) => {
+	const styles = useContext(StylesContext);
+
 	return (
-		<div className='form-group'>
+		<div className={`${styles['form-group']} ${styles['form-input']}`}>
 			<label htmlFor={name}>{label}</label>
 			<input
 				type={type}
@@ -10,7 +13,7 @@ export const FormInput = ({ label, type, name, value, onChange }) => {
 				id={name}
 				value={value}
 				onChange={onChange}
-				className='form-control'
+				className={styles['form-control']}
 			/>
 		</div>
 	);
