+++ src/components/Auth/FormComponents/FormInput.module.css
@@ -1,0 +1,14 @@
+.form-group {
+	margin-bottom: 20px;
+}
+
+.form-group label {
+	font-weight: bold;
+}
+
+.form-control {
+	width: 100%;
+	padding: 10px;
+	border: 2px solid #ccc;
+	border-radius: 4px;
+}
