+++ src/components/Auth/FormComponents/ThemeContext.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const ThemeContext = React.createContext();
+
+export default ThemeContext;
