+++ src/components/Auth/Login/Login.module.css
@@ -42,3 +42,18 @@
 .auth-container form button:hover {
 	background-color: #3e8e41;
 }
+
+.form-group {
+	margin-bottom: 20px;
+}
+
+.form-group label {
+	font-weight: bold;
+}
+
+.form-control {
+	width: 100%;
+	padding: 10px;
+	border: 2px solid #ccc;
+	border-radius: 4px;
+}
